Email-friendly handling of newlines in blockquotes
Multiple quoting levels are common in emails and they are rarely
separated by blank lines. This is not a problem with increasing
levels, but it's a problem with decreasing levels when so-called
lazy continuation lines come into play. When CMARK_OPT_HARDBREAKS
is enabled then this behavior is now disabled. It should perhaps
have its own option value.